packages/ryuuko-api/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"ryuuko_api" = "src"}
packages = ["ryuuko_api"]

[project]
name = "ryuuko-api"
version = "2.0.1" # Incrementing version
authors = [
  { name="Zang Vũ", email="zvwgvx@polydevs.uk" },
]
description = "Core API Service for the Ryuuko Chatbot ecosystem."
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: 3",
    "Framework :: FastAPI",
    "Operating System :: OS Independent",
]
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "python-dotenv>=1.0.0",
    "pymongo>=4.0.0",
    "google-generativeai>=0.5.0",
    "openai>=1.3.0",
    # --- Added for Dashboard Auth ---
    "pydantic[email]",
    "passlib>=1.7.4",
    "bcrypt==3.1.7", # Pinning bcrypt to a compatible version
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.9"
]

[project.scripts]
ryuuko-api = "ryuuko_api.runner:main"
